refac: move line type tests to different module
